#84 Streak Powerup (#160)

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/model/game/Game.java
@@ -16,7 +16,7 @@ public class Game {
 
     private Long activePlayer;
 
-    private Long activePlayerStreak;
+    private Integer activePlayerStreak;
 
     private GameState gameState;
